backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.25.0
requests==2.31.0
gitpython==3.1.40
groq==0.4.1